+++ src/RestApi/Alerts/IAlertFilter.ts
@@ -1,0 +1,7 @@
+export default interface IAlertFilter {
+	pagination?: string;
+	createdUntil?: string;
+	until?: Date;
+	archived?: boolean;
+	snoozed?: boolean;
+}