+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Location `create`, `list` (multiple), `get` (single), `update` and `delete`
 - Organization tag to location `assign` and `unassign`
 - Location to device `assign` and `unassign`
+- Alert and alert rules management
 
 ## [1.7.1] - 2021-12-17
 ### Fixed